refactor(js): consolidate xterm hooks into a single shared factory
Both terminal_panel.js and service_log_viewer.js shared the same xterm
setup (themes, fit addon, theme listener, resize observer, cleanup) with
small behavioral differences. Replace them with a single xterm_hook.js
exposing a createXtermHook(opts) factory and two named exports
(TerminalPanel, ServiceLogViewer) tuned for interactive vs read-only use.
