+++ assets/js/app.js
@@ -24,8 +24,7 @@ import {Socket} from "phoenix"
 import {LiveSocket} from "phoenix_live_view"
 import {hooks as colocatedHooks} from "phoenix-colocated/destila"
 import topbar from "../vendor/topbar"
-import TerminalPanel from "./hooks/terminal_panel"
-import ServiceLogViewer from "./hooks/service_log_viewer"
+import {TerminalPanel, ServiceLogViewer} from "./hooks/xterm_hook"
 import LocalTime from "./hooks/local_time"
 import DraftsBoard from "./hooks/drafts_board"
